fix products

+++ GPA.Dtos/Inventory/ProductCreationDto.cs
@@ -9,7 +9,7 @@
         public decimal Price { get; set; }
         public string Description { get; set; }
         public byte Type { get; set; }
-        public int UnitValue { get; set; }
+        public double UnitValue { get; set; }
         public Guid UnitId { get; set; }
         public Guid CategoryId { get; set; }
         public Guid? ProductLocationId { get; set; }
